requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
apscheduler==3.11.0
beautifulsoup4==4.13.4
requests==2.32.3
python-docx==1.1.2
pypdf==5.4.0
markdown==3.8
pyyaml==6.0.2
fpdf2==2.8.2
